Initial repository creation.
Created a structure corresponding to a complex number; the members are
of double precision. (We do not use the GNU intrinsic complex arithmetic
operations to preserve code portability.)

Added the most essential operators for the elements of the field of complex
numbers, with some special ones for multiplication by pure imaginary and
pure real elements. (The set of operators is not complete, but this is all
I need for what I am doing now.)
IN 2019/06/07